requirements.txt
allure-pytest==2.13.1
allure-python-commons===2.13.1
appdirs==1.4.4
astroid==2.3.3
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.8.1
browserstack-local==1.2.2
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
ConfigArgParse==1.5.3
confluent-kafka==1.4.0
cryptography==38.0.1
cycler==0.11.0
deepdiff==5.0.2
dictor==0.1.5
execnet==1.9.0
Faker==2.0.3
fonttools==4.33.3
google-api-python-client==1.7.11
google-auth==2.9.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
greenlet==0.4.16
h11==0.13.0
httplib2==0.20.4
idna==2.8
imageio==2.19.2
importlib-metadata==4.11.4
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.2
jsonlines==1.2.0
kiwisolver==1.4.2
kubernetes==11.0.0
lazy-object-proxy==1.4.3
locustio==0.14.6
lorem-text==1.5
MarkupSafe==2.1.1
matplotlib==3.5.2
mccabe==0.6.1
more-itertools==8.13.0
msgpack==1.0.3
networkx==2.8.2
numpy==1.22.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
ordered-set==4.1.0
outcome==1.1.0
packaging==21.3
pandas==1.0.5
Pillow==9.3.0
pluggy==0.13.1
psutil==5.9.1
psycopg2-binary==2.8.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pylint==2.4.4
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==5.0.1
pytest-dependency==0.5.1
pytest-forked==1.4.0
pytest-html==1.22.0
pytest-json-report==1.2.4
pytest-metadata==2.0.1
pytest-parallel==0.0.9
pytest-rerunfailures==7.0
pytest-xdist==1.30.0
python-dateutil==2.8.2
python-decouple==3.1
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.7
scikit-image==0.19.2
scipy==1.10.0
selenium==4.9.1
blinker==1.6.2
selenium-wire==5.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.17
str2bool==1.1
text-unidecode==1.3
trio==0.20.0
trio-websocket==0.9.2
uritemplate==3.0.1
uritools==4.0.0
urlextract==0.14.0
urllib3==1.26.13
wcwidth==0.2.5
websocket-client==1.3.2
Werkzeug==2.2.3
wrapt==1.11.2
wsproto==1.1.0
xlrd==1.2.0
zipp==3.8.0
configparser~=5.2.0
gevent~=20.4.0
elasticsearch==7.13.4
openpyxl==3.1.2
